requirements.txt
absl-py
adal
addict
aiohttp
analytics-python
anyio
applicationinsights
appnope
argcomplete
argon2-cffi
arrow
asgiref
astor
astunparse
async-generator
async-timeout
attrs
azure-common
azure-core
azure-graphrbac
azure-identity
azure-mgmt-authorization
azure-mgmt-containerregistry
azure-mgmt-core
azure-mgmt-keyvault
azure-mgmt-resource
azure-mgmt-storage
azure-storage-blob
azureml
azureml-core
azureml-dataprep
azureml-dataprep-native
azureml-dataprep-rslex
azureml-dataset-runtime
azureml-telemetry
azureml-tensorboard
azureml-train
azureml-train-core
azureml-train-restclients-hyperdrive
azureml-widgets
Babel
backcall
backports.tempfile
backports.weakref
bleach
blinker
boto3
botocore
brotlipy
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
commonmark
configparser
contextlib2
coolname
coverage
cryptography
cycler
Cython
dataclasses-json
debugpy
decorator
defusedxml
Deprecated
deprecation
distro
docker
docker-pycreds
docopt
docutils
dotnetcore2
entrypoints
falcon
fastapi
flatbuffers
fsspec
future
fvcore
gast
gitdb
GitPython
google-api-core
google-auth
google-auth-oauthlib
google-cloud
google-cloud-core
google-cloud-storage
google-pasta
google-resumable-media
googleapis-common-protos
GPUtil
gql
graphql-core
grpcio
gunicorn
h11
h5py
humanize
idna
importlib-metadata
iopath
ipykernel
ipython
ipython-genutils
ipywidgets
isodate
jedi
jeepney
Jinja2
jmespath
joblib
json5
jsonpickle
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyter-packaging
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
Keras-Preprocessing
kiwisolver
lightning-grid
Markdown
MarkupSafe
marshmallow
marshmallow-enum
matplotlib
matplotlib-inline
maxminddb
mistune
mkl-fft
mkl-random
mkl-service
more-itertools
msal
msal-extensions
msrest
msrestazure
multidict
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
ndg-httpsclient
nest-asyncio
notebook
numpy
oauthlib
olefile
opencv-python
opencv-python-headless
opt-einsum
packaging
pandas
pandocfilters
parso
pathspec
pathtools
pexpect
pickleshare
Pillow
pip
pipreqs
pluggy
plyfile
portalocker
pprintpp
prometheus-client
promise
prompt-toolkit
protobuf
psutil
ptyprocess
py
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pyDeprecate
PyGithub
Pygments
PyJWT
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil
python-multipart
pytorch-lightning
pytz
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
requests-unixsocket
rich
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
SecretStorage
Send2Trash
sentry-sdk
setuptools
shortuuid
sip
six
smmap
sniffio
starlette
stringcase
subprocess32
tabulate
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
tflite-runtime
threadpoolctl
tomlkit
torch
torchmetrics
torchvision
tornado
tqdm
traitlets
typing_extensions
typing-inspect
tzlocal
ulid-py
urllib3
uvicorn
wandb
wcwidth
webencodings
websocket-client
websockets
Werkzeug
wheel
widgetsnbextension
wrapt
yacs
yarg
yarl
yaspin
zipp
